+++ testing/src/main/scala/regressionTests.scala
@@ -23,7 +23,7 @@ import gapt.provers.congruence.SimpleSmtSolver
 import gapt.provers.escargot.Escargot
 import gapt.provers.prover9.Prover9Importer
 import gapt.provers.sat.{ MiniSAT, Sat4j }
-import gapt.provers.smtlib.Z3
+import gapt.provers.smtlib.{ SmtInterpol, Z3 }
 import gapt.provers.verit.VeriT
 import gapt.provers.viper.grammars.EnumeratingInstanceGenerator
 import gapt.provers.viper.{ Viper, ViperOptions }
@@ -210,6 +210,7 @@ class Prover9TestCase( f: java.io.File ) extends RegressionTestCase( f.getParent
     }
     ExpansionProofToLK( E ).isRight !-- "expansionProofToLKProof"
     Z3.isValid( deep ) !-- "validity of deep formula"
+    SmtInterpol.isValid( deep ) !-- "validity of deep formula (SmtInterpol)"
     SimpleSmtSolver.isValid( deep ) !-- "SimpleSmtSolver on deep formula"
 
     if ( isFOLPrenexSigma1( p.endSequent ) )